Moved "tasks" to entry point
